CSharpDemo: increased integration resolution to µs

+++ sample-code/CSharpDemo/Spectrometer.cs
@@ -21,6 +21,6 @@ interface Spectrometer {
 	int[] getEDCIndices();
 	double[] getNLCCoeffs();
 
-	bool setIntegrationTimeMilliseconds(int ms);
+	bool setIntegrationTimeMilliseconds(double ms);
 	bool setTriggerMode(int mode);
 }
